Convert heel options to permit double values (to match actuas).
BugzId: 1131

+++ com-mmxlabs-models-lng/com.mmxlabs.models.lng.fleet.editor/src/com/mmxlabs/models/lng/fleet/presentation/composites/HeelOptionsInlineEditorWrapper.java
@@ -32,8 +32,8 @@ public class HeelOptionsInlineEditorWrapper extends IInlineEditorEnablementWrapp
 	@Override
 	protected boolean isEnabled() {
 		HeelOptions options = (HeelOptions) input;
-		Integer vol = options.getVolumeAvailable();
-		return vol > 0;
+		Double vol = options.getVolumeAvailable();
+		return vol > 0.0;
 	}
 
 	@Override
